sidebar: strip sidebar update functions
Remove the hard-coded sidebar update functions.
The sidebar now relies on notifications / observer functions.
